Move federation rest code under relapse.rest
