+++ relapse/rest/admin/federation.py
@@ -17,10 +17,10 @@ from typing import TYPE_CHECKING
 
 from relapse.api.constants import Direction
 from relapse.api.errors import Codes, NotFoundError, RelapseError
-from relapse.federation.transport.server import Authenticator
 from relapse.http.servlet import RestServlet, parse_enum, parse_integer, parse_string
 from relapse.http.site import RelapseRequest
 from relapse.rest.admin._base import admin_patterns, assert_requester_is_admin
+from relapse.rest.federation import Authenticator
 from relapse.storage.databases.main.transactions import DestinationSortOrder
 from relapse.types import JsonDict
